react-project/backend/requirements.txt
Flask
python-dotenv
Flask-PyMongo
Flask-JWT-Extended
Flask-Cors
Werkzeug
